requirements.txt
numpy
opencv-python
matplotlib
Pillow
kagglehub
